[Fix] users api team으로 이동
