--- frontend/src/apis/users.ts
@@ -1,13 +1,0 @@
-import fetchApi from '@utils/fetch';
-
-export const readTeamInfo = async (id: number) => {
-	const res = await fetchApi.get(`/api/team/${id}`);
-	const data = await res.json();
-	return data;
-};
-
-export const readTeamUsers = async (id: number) => {
-	const res = await fetchApi.get(`/api/team/users/${id}`);
-	const data = await res.json();
-	return data;
-};
